[refactor] website: rename 1-1 advising slug to /programs/advising
