+++ apps/website/src/pages/programs.tsx
@@ -59,10 +59,10 @@ const ProgramsPage = () => {
       ctaLabel: 'Explore program',
     },
     {
-      id: '1-1-advising',
+      id: 'advising',
       name: '1-1 advising',
       status: 'Active',
-      href: '/programs/1-1-advising',
+      href: '/programs/advising',
       summary: 'A 20-min call with the BlueDot team to accelerate you towards impactful work in AI safety.',
       detail: '200+ advising calls done. Decisions in ~5 working days.',
       ctaLabel: 'Explore program',
